Extract frozen device parsing utils to separate file
